WDL updates, templates
Added functionality for AWS, including inputs to WDL that take in AWS information and a Docker image. Also, removed extra output file from format_taxa and rely on glob returning files in order. Made templates for JSON file for input into WDL and configuration for a Cromwell 42 server on AWS